delete機能とupdate機能の改修

+++ novelSite/src/main/java/io/post/novel/controller/UserDisplayController.java
@@ -128,43 +128,113 @@ public class UserDisplayController {
 	@PostMapping("/user/password/edit/check")
 	public String passwordCheck(@ModelAttribute @Validated PassValid passValid, BindingResult result, Model model, @AuthenticationPrincipal UserForm userForm) {
 		
+		model.addAttribute("pass_valid", passValid);
 		
 		if (result.hasErrors()) {
-			 	model.addAttribute("pass_valid", passValid);
+			 	
 			  
 			  return "user/edit_password";//エラー時は自画面遷移
 	        }
-		//passValid.setId(userForm.getId());
-		//userDisplayService.findUserOne(passValid);
 		
+		//入力されたパスワードとセッションのパスワードの一致確認
 		boolean passCheck = encoder.matches(passValid.getPassword(), userForm.getPassword());
-		System.out.println(passCheck);
 		
-		return "redirect:/user/user_page";
+		if(passCheck == true) {
+			
+			return "user/input_new_password";
+		}
+		
+		return "user/edit_password";//一致しない場合は自画面繊維
 	}
 	
+	
+	/*
+	 * パスワード変更処理実行
+	 */
+	@PostMapping("/user/change/password")
+	public String passChange(@ModelAttribute @Validated PassValid passValid, BindingResult result, Model model, @AuthenticationPrincipal UserForm userForm) {
+		
+		model.addAttribute("change_pass", passValid);
+		
+		//レコード指定用のidをprincipalから取得
+		//取得したidをpassValidにセット
+		long id = userForm.getId();
+		passValid.setId(id);
+		
+		if (result.hasErrors()) {
+		 	
+		  return "user/input_new_password";//エラー時は自画面遷移
+        }
+		
+		//入力で受け取ったパスワードのハッシュ化
+		String rawPassword = passValid.getPassword();
+		passValid.setPassword(encoder.encode(rawPassword));
+		
+		//DBのデータをUpdate処理
+		userDisplayService.updatePass(passValid);
+		
+		//ハッシュ化したパスワードをPrincipalにセット
+		userForm.setPassword(passValid.getPassword());
+		
+		//認証情報の「ゆーざーねーむぱすわーどおうせんてぃけーしょんとーくん」を更新
+		 SecurityContext auth = SecurityContextHolder.getContext();
+		 auth.setAuthentication(new UsernamePasswordAuthenticationToken(userForm, userForm.getPassword(), userForm.getAuthorities()));
+		
+		return "redirect:/user/user_page";
+		
+	}
 
 	
+	/*
+	 * デリート画面へ遷移
+	 */
 	@GetMapping("/user/user_delete")
 	public String toDeleteDisplay() {
 		
 		return "user/delete";
 	}
 	
+	
+	/*
+	 * デリートユーザー取得
+	 */
 	@PostMapping("/user/user_delete/check")
-	public String deleteInfoCheck(@ModelAttribute UserRequest deleteUser, Model model) {
+	public String deleteInfoCheck(@ModelAttribute UserRequest deleteUser, Model model,@AuthenticationPrincipal UserForm userForm) {
 		
+		
+		//入力されたパスワードとセッションのパスワードの一致確認
+		boolean passCheck = encoder.matches(deleteUser.getPassword(), userForm.getPassword());
+		
+		//入力で受け取ったパスワードのハッシュ化
+		String rawPassword = deleteUser.getPassword();
+		deleteUser.setPassword(encoder.encode(rawPassword));
+		
+		//入力内容と一致しない場合は自画面遷移
+		if(passCheck != true) {
+					
+			return "user/delete";
+		}
+		
+		//セッションからidを取得してdeleteUserにセット
+		deleteUser.setId(userForm.getId());
+		
+		//削除対象のユーザー情報をDBから取得
 		SignUpUser deleteUserInfo = userDisplayService.deleteUserSelect(deleteUser);
+		//取得内容をmodelに詰めて画面へ表示
 		model.addAttribute("delete_user", deleteUserInfo);
 		
 		return "user/delete_check";
 	}
 	
+	/*
+	 * デリート処理実行
+	 */
 	@PostMapping("/user/user_delete/check/{id}")
-	public String delete(@PathVariable("id") String id) {
-		
+	public String delete(@ModelAttribute UserRequest delete, @PathVariable("id") String id) {
+		//pathでわたってきたidを使ってユーザーデリート実行
 		userDisplayService.delete(id);
 		
-		return "redirect:/top";
+		//リダイレクトでログアウト
+		return "redirect:/signout";
 	}
  }
